fix(oathkeeper): remove default configmap
The following empty config map is created because since we don’t have any ory/rules in the cluster, the resource is not created.

But when we apply the Kustomize it erase the content written by the controller. And it will not be override by them until a new ory/rule is changed.

This will fail the deployment of oathkeeper since a ory/rule is added.

Signed-off-by: iverly <github@iverly.net>